+++ lib/src/extensions/code_extension.dart
@@ -1,0 +1,25 @@
+part of '../../flutter_supabase_macro.dart';
+
+extension on Code {
+  /// Used for error messages.
+  String get debugString {
+    final buffer = StringBuffer();
+    _writeDebugString(buffer);
+    return buffer.toString();
+  }
+
+  void _writeDebugString(StringBuffer buffer) {
+    for (final part in parts) {
+      switch (part) {
+        case Code():
+          part._writeDebugString(buffer);
+        case Identifier():
+          buffer.write(part.name);
+        case OmittedTypeAnnotation():
+          buffer.write('<omitted>');
+        default:
+          buffer.write(part);
+      }
+    }
+  }
+}