+++ lib/src/extensions/iterable_extension.dart
@@ -1,0 +1,10 @@
+part of '../../flutter_supabase_macro.dart';
+
+extension _IterableExtension<T> on Iterable<T> {
+  T? firstWhereOrNull(bool Function(T) compare) {
+    for (final item in this) {
+      if (compare(item)) return item;
+    }
+    return null;
+  }
+}