+++ lib/src/extensions/type_declaration_extension.dart
@@ -1,0 +1,7 @@
+part of '../../flutter_supabase_macro.dart';
+
+extension _IsExactly on TypeDeclaration {
+  /// Cheaper than checking types using a [StaticType].
+  bool isExactly(String name, Uri library) =>
+      identifier.name == name && this.library.uri == library;
+}