+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## 0.0.1
 
 Initial release : 
-- Creation of a `toJsonSupabase` which exclude the primary key from the `Map`
+- Creation of a `toJsonSupabase` which exclude the `primaryKey` from the `Map`